[Oztechan/CCC#3803] Clean Application class and move view related things into MainView
